// src/app/models/machine.model.ts
export interface Machine {
  id: String;
  assetcode: String;
  asset: String;
  trademark: String;
  code: String;
  location: String;
  model: String;
  series: String;
  voltage: String;
  measurement: String;
  power: String;
  inventory: String;
  speed: String;
  reference: String;
  capacity: String;
}